backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
firebase-admin
pydantic
google-genai
python-dotenv
Pillow
